% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karlen-package.R
\docType{package}
\name{karlen-package}
\alias{karlen}
\alias{karlen-package}
\title{karlen: Real-Time PCR Data Sets by Karlen et al. (2007)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
}

}
\keyword{internal}
